Move xDai claim data to your connection modal
